FIXED JonWorld5 and JonWorld6
I accidentally had a portion that was supposed to be marked negative marked positive. Luckily it was really easy to notice because I had it setup so the positives were first then the negatives, and I realized I had a lot of positives in the negative section. They are now (correctly) negative.